backend/requirements.txt
Flask>=3.0.0
Flask-CORS>=4.0.0
Flask-SocketIO>=5.3.0
google-cloud-speech>=2.33.0
google-cloud-storage>=3.3.0
langchain-core>=0.3.75
langchain-google-genai>=2.1.10
numpy>=2.0.2
python-dotenv>=1.1.1
scipy>=1.11.0
soundfile>=0.13.1
gunicorn>=21.2.0
python-socketio>=5.10.0
eventlet>=0.33.0
pydub>=0.25.1
